Fixes several bugs mostly in stratified_split
Tests for stratified split where bugged and did not catch wrong lenghts,
because any object evaluates to True. Now use of any fixes this.

Added additional Tests for edgecases.

Fixed the behaviour of stratified split. Now it respects the given
lenghts. Additionally it now reports an error if the size of
the desired subsets is to small given the number of classes and minimum
samples.