Fixes a bug where labels where randomly swapped
The Taskdataloader object contained an error which would load the wrong data for the associated labels because when it collected the indices from the wrong
list. This is corrected by using the correct list. Additionally this change needs to sort sample_ids now because they are in a random order, which hdf5 does not support.