Adds a new blacklist functionality
The blacklist functionality is intended to avoid reusing samples from the training set
in the testing set. Therefore we use the meta_train flag to write the set of blacklisted items
once we iterate over the meta_train set. When finished we can pass the blacklist to a meta_testset where
the flag is False. Those samples are then excluded. To this end there is also a max_samples_per_task flag
necessary in TCGA to not block all samples.